+++ HISTORY.rst
@@ -13,6 +13,7 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
 **Bug fixes**
 
 - #714: [OpenBSD] virtual_memory().cached value was always set to 0.
+- #715: don't crash at import time if cpu_times() fail for some reason.
 
 
 3.3.0 - 2015-11-25